# Use codegen instead of reflect for NodeEqual