… src/php/astutil/equal.go
… src/php/astutil/gen_equal.go
… src/solver/node_equal.go