Our current NodeEqual from solver package will panic on
slices that are not []node.Node. Previously all slices
were like that, but now we have several specialized
slices (e.g. []*node.Identifier).

Another problem is that it's quite slow. We can't use it
in algorithms that require a lot of comparisons. One
example is code duplication detection.

Using code generation solves both correctness and performance issues.

"go generate" can be used to re-generate NodeEqual code:

        go generate ./src/...

----

Benchstat results (reflect vs codegen):

```
name            old time/op    new time/op    delta
ReflectEqual-8    4.38µs ± 0%    0.08µs ± 0%   -98.12%  (p=0.016 n=4+5)

name            old alloc/op   new alloc/op   delta
ReflectEqual-8     96.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)

name            old allocs/op  new allocs/op  delta
ReflectEqual-8      6.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
```

----

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>